requirements.txt
openai>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
pydantic>=2.4.0
pytest>=7.4.0
seaborn>=0.12.0 